requirements.txt
numpy
pandas
faker
### datetime
mysql.connector-python